Refactor jetty server handler
